requirements.txt
Flask
lxml
pyjwt
signxml
